% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-data.R
\docType{data}
\name{avengers}
\alias{avengers}
\alias{avengers_wide}
\title{Power ratings for The Avengers.}
\format{A data frame with 24 rows and 4 variables:
\describe{
  \item{group}{Name of the hero}
  \item{axis}{Power skill}
  \item{value}{Value (1-7)}
  \item{description}{Character description}
}}
\source{
Marvel Wikia (\url{http://marvel.wikia.com}) and Chris Zhou (\url{http://bl.ocks.org/chrisrzhou/2421ac6541b68c1680f8})
}
\usage{
avengers

avengers_wide
}
\description{
Data are available in "long" and "wide" format.
}
\keyword{datasets}
