% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_area}
\alias{bb_area}
\title{Area property for a Billboard.js chart}
\usage{
bb_area(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.area>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Area property for a Billboard.js chart
}
\examples{
# Area chart options
billboarder() \%>\%
  bb_linechart(data = c(2, 4, 3, 6, 5), type = "area") \%>\%
  bb_area(linearGradient = TRUE)
}
