% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_click}
\alias{bb_click}
\title{Retrieve click value in Shiny}
\usage{
bb_click(bb, inputId)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{inputId}{The \code{input} slot that will be used to access the value.}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Retrieve click value in Shiny
}
\examples{
\dontrun{

 if (interactive()) {
 
 
 }

}
}
