% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_densityplot}
\alias{bb_densityplot}
\title{Helper for creating a density plot}
\usage{
bb_densityplot(
  bb,
  data,
  mapping = NULL,
  stacked = FALSE,
  stat = "density",
  fill = FALSE,
  ...
)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{data}{A `data.frame` or a `vector`, the first column will
be used to calculate density if `x` is `NULL`.}

\item{mapping}{Mapping of variables on the chart, see [bbaes()].}

\item{stacked}{Logical, create a stacked density plot.}

\item{stat}{Stat to compute: `density` or `count`.}

\item{fill}{Produce a conditional density estimate, this option force `stacked = TRUE`.}

\item{...}{Arguments passed to [stats::density()].}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Helper for creating a density plot
}
\examples{
# With a vector
billboarder() \%>\%
  bb_densityplot(data = rnorm(1e4))

data("diamonds", package = "ggplot2")

# density plot with one variable
billboarder() \%>\%
  bb_densityplot(data = diamonds, x = "carat")

# Same with mapping
billboarder() \%>\%
  bb_densityplot(diamonds, bbaes(carat))

# With a grouping variable
billboarder() \%>\%
  bb_densityplot(data = diamonds, x = "depth", group = "cut") \%>\%
  bb_x_axis(min = 55, max = 70)

# Same with mapping
billboarder() \%>\%
  bb_densityplot(diamonds, bbaes(depth, group = cut)) \%>\%
  bb_x_axis(min = 55, max = 70)

# a stacked density plot using count as statistic
bb <- billboarder() \%>\%
  bb_densityplot(
    diamonds, bbaes(depth, group = cut),
    stacked = TRUE, stat = "count"
  ) \%>\%
  bb_x_axis(min = 55, max = 70)
bb

# changing order
bb \%>\% bb_data(order = "asc")
}
\seealso{
[bb_histogram()]
}
