% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_donut}
\alias{bb_donut}
\title{Donut property for a Billboard.js chart}
\usage{
bb_donut(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.donut>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Donut property for a Billboard.js chart
}
\examples{
billboarder() \%>\%
  bb_donutchart(data = table(mtcars$cyl)) \%>\%
  bb_donut(title = "Donut Title", width = 10)
}
