% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_hist}
\alias{bb_hist}
\title{Helper for creating an histogram}
\usage{
bb_hist(bb, x, breaks = "Sturges", ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{Arguments for slot}

\item{data}{A \code{data.frame}.}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating an histogram
}
\examples{
\dontrun{
}
}
