% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_line}
\alias{bb_line}
\title{Line property for a Billboard.js chart}
\usage{
bb_line(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.line>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Line property for a Billboard.js chart
}
\examples{
# Set if null data point will be connected or not.
b <- billboarder() \%>\%
  bb_linechart(data = c(1, 2, NA, 4, 5))
b
b \%>\% bb_line(connectNull = TRUE)
}
