% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_load}
\alias{bb_load}
\title{Load data to the chart with proxy}
\usage{
bb_load(proxy, data = NULL, unload = NULL, ...)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} with updated data.}

\item{unload}{Ids (names) to data to unload.}

\item{...}{Arguments passed to method.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Load data to the chart with proxy
}
