% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_padding}
\alias{bb_padding}
\title{The padding of the chart element.}
\usage{
bb_padding(bb, ...)
}
\arguments{
\item{bb}{A [billboarder()] `htmlwidget` object or a [billboarderProxy()] `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.padding> for possible options.}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
The padding of the chart element.
}
\examples{
# Add padding around the chart
billboarder() \%>\%
  bb_barchart(data = table(sample(letters[1:5], 100, TRUE))) \%>\%
  bb_padding(top = 100, right = 100, bottom = 100, left = 100)
}
