% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_point}
\alias{bb_point}
\title{Point property for a Billboard.js chart}
\usage{
bb_point(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.point>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Point property for a Billboard.js chart
}
\examples{
# Set point size
billboarder() \%>\%
  bb_scatterplot(data = iris, x = "Sepal.Length", y = "Sepal.Width", group = "Species") \%>\%
  bb_axis(x = list(tick = list(fit = FALSE))) \%>\%
  bb_point(r = 10)
}
