% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_axis_labels}
\alias{bb_proxy_axis_labels}
\title{Update axis labels with proxy}
\usage{
bb_proxy_axis_labels(proxy, x = NULL, y = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{x}{X axis label.}

\item{y}{Y axis label.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Update axis labels with proxy
}
