% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_flow}
\alias{bb_proxy_flow}
\title{Update chart flow with proxy}
\usage{
bb_proxy_flow(proxy, ...)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{...}{Arguments passed to the flow API, see \url{https://naver.github.io/billboard.js/release/latest/doc/Chart.html#flow}.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Update chart flow with proxy
}
