% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_groups}
\alias{bb_proxy_groups}
\title{Update chart groups with proxy}
\usage{
bb_proxy_groups(proxy, ...)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{...}{Vector(s) with id of the series, e.g. the name of variables.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Update chart groups with proxy
}
