% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_hide}
\alias{bb_proxy_hide}
\title{Hide method with proxy}
\usage{
bb_proxy_hide(proxy, targetIdsValue, options = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{targetIdsValue}{Name of series to hide.}

\item{options}{Additional options.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Hide method with proxy
}
\seealso{
\code{\link{bb_proxy_show}}
}
