% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_legend}
\alias{bb_proxy_legend}
\title{Show or hide legend with proxy}
\usage{
bb_proxy_legend(proxy, what = c("show", "hide"), targetIds = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{what}{\code{show} or \code{hide} the legend.}

\item{targetIds}{Series ids to show/hide, if \code{NULL} show/hide all legend.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Show or hide legend with proxy
}
