% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_show}
\alias{bb_proxy_show}
\title{Show method with proxy}
\usage{
bb_proxy_show(proxy, targetIdsValue, options = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{targetIdsValue}{Name of series to show.}

\item{options}{Additional options.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Show method with proxy
}
\seealso{
\code{\link{bb_proxy_hide}}
}
