% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_tooltip}
\alias{bb_proxy_tooltip}
\title{Show or hide tooltip with proxy}
\usage{
bb_proxy_tooltip(proxy, what = c("show", "hide"), x = NULL, index = NULL, ...)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{what}{\code{show} or \code{hide} the legend.}

\item{x}{x value on which the tooltip must appear.}

\item{index}{Index on the x-axis on which the tooltip must appear.}

\item{...}{Additional arguments passed to method.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Show or hide tooltip with proxy
}
