% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_transform}
\alias{bb_proxy_transform}
\title{Update chart type with proxy}
\usage{
bb_proxy_transform(proxy, type, targetIds = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{type}{Specify the type to be transformed.}

\item{targetIds}{Specify targets to be transformed. If not given, all targets will be the candidate.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Update chart type with proxy
}
