% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_proxy_xs}
\alias{bb_proxy_xs}
\title{Update x values with proxy}
\usage{
bb_proxy_xs(proxy, xs)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{xs}{Named list of vector(s) used for x values.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Update x values with proxy
}
