% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_radarchart}
\alias{bb_radarchart}
\title{Helper for creating a radar chart}
\usage{
bb_radarchart(bb, data, mapping = NULL, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{data}{A `data.frame`, the first column will be used for x axis unless
specified otherwise in `mapping`. If not a `data.frame`, an object coercible to `data.frame`.}

\item{mapping}{Mapping of variables on the chart, see [bbaes()].}

\item{...}{Arguments passed to [bb_radar()].}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Helper for creating a radar chart
}
\examples{
library("billboarder")

# data about Avengers
data("avengers_wide")

# if not specified, first column is used as x-axis,
# all others are used on y-axis
billboarder() \%>\%
  bb_radarchart(data = avengers_wide)

# specify explicitly which column to use with mapping
billboarder() \%>\%
  bb_radarchart(
    data = avengers_wide,
    mapping = bbaes(x = axis, y = `Captain America`)
  )

# with data in "long" format you can use "group" aesthetics
data("avengers")
billboarder() \%>\%
  bb_radarchart(
    data = avengers,
    mapping = bbaes(x = axis, y = value, group = group)
  )
}
