% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_render}
\alias{bb_render}
\title{Render property for a Billboard.js chart}
\usage{
bb_render(bb, ...)
}
\arguments{
\item{bb}{A [billboarder()] `htmlwidget` object or a [billboarderProxy()] `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.render> for possible options.}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Render property for a Billboard.js chart
}
