% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_scatterplot}
\alias{bb_scatterplot}
\title{Helper for creating a scatter chart}
\usage{
bb_scatterplot(bb, data, mapping = NULL, ..., point_opacity = NULL)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{data}{A `data.frame`}

\item{mapping}{Mapping of variables on the chart, see [bbaes()].}

\item{...}{Alternative mapping, you can specify `x = "Sepal.Length"` for example.}

\item{point_opacity}{Opacity for points, value between `[0,1]`.}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Helper for creating a scatter chart
}
\note{
This function can be used with [billboarderProxy()] in Shiny applications.
}
\examples{
# Use first and second variable by default
billboarder() \%>\%
  bb_scatterplot(data = iris)

# Explicit mapping
billboarder() \%>\%
  bb_scatterplot(
    data = iris,
    mapping = bbaes(Petal.Length, Petal.Width)
  ) \%>\%
  bb_x_axis(tick = list(fit = FALSE))

# Grouping variable
billboarder() \%>\%
  bb_scatterplot(
    data = iris,
    mapping = bbaes(Sepal.Length, Sepal.Width, group = Species)
  )

# Size variable
billboarder() \%>\%
  bb_scatterplot(
    data = iris,
    mapping = bbaes(
      Sepal.Length, Sepal.Width,
      group = Species, size = Petal.Width
    )
  ) \%>\%
  bb_x_axis(tick = list(fit = FALSE))
}
