% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_spline}
\alias{bb_spline}
\title{Spline property for a Billboard.js chart}
\usage{
bb_spline(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.spline>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Spline property for a Billboard.js chart
}
\examples{
# Customize spline interpolation
billboarder() \%>\%
  bb_linechart(data = c(1, 5, 3, 6, 2), type = "spline") \%>\%
  bb_spline(interpolation = list(type = "natural"))
}
