% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_svg}
\alias{bb_svg}
\title{SVG property for a Billboard.js chart}
\usage{
bb_svg(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.svg>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
SVG property for a Billboard.js chart
}
\examples{
# Set SVG class name
billboarder() \%>\%
  bb_linechart(data = c(2, 5, 3, 4, 6)) \%>\%
  bb_svg(classname = "custom-billboard")
}
