% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_title}
\alias{bb_title}
\title{Add title to Billboard.js chart}
\usage{
bb_title(bb, text = NULL, padding = NULL, position = "top-center", ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{text}{The chart title.}

\item{padding}{A named list with `top`, `right`, `bottom`, `left` values.}

\item{position}{A string specifying the position of the title.}

\item{...}{Additional arguments.}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Add title to Billboard.js chart
}
\examples{
billboarder() \%>\%
  bb_barchart(data = table(sample(letters, 100, TRUE))) \%>\%
  bb_title(text = "Random letters", position = "center")
}
\seealso{
[bb_labs()]
}
