% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_transition}
\alias{bb_transition}
\title{Transition property for a Billboard.js chart}
\usage{
bb_transition(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.transition>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Transition property for a Billboard.js chart
}
\examples{
# Set transition duration
billboarder() \%>\%
  bb_linechart(data = c(5, 3, 6, 2, 7)) \%>\%
  bb_transition(duration = 1000)
}
