% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_treemapchart}
\alias{bb_treemapchart}
\title{Helper for creating a treemap chart}
\usage{
bb_treemapchart(bb, data, mapping = NULL, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{data}{A `data.frame`, the first column will be used for x axis unless
specified otherwise in `mapping`. If not a `data.frame`, an object coercible to `data.frame`.}

\item{mapping}{Mapping of variables on the chart, see [bbaes()].}

\item{...}{Arguments passed to [bb_treemap()].}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Helper for creating a treemap chart
}
\examples{
library("billboarder")
data("mpg", package = "ggplot2")

billboarder() \%>\%
  bb_treemapchart(mpg[, 1])

billboarder() \%>\%
  bb_treemapchart(
    data = mpg,
    mapping = aes(x = manufacturer),
    label = list(show = TRUE, threshold = 0.3)
  ) \%>\%
  bb_data(
    labels = list(colors = "#FFF")
  )
}
