% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_unload}
\alias{bb_unload}
\title{Unload data to the chart with proxy}
\usage{
bb_unload(proxy, ids = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{ids}{Data ids to unload.}
}
\value{
A \code{billboardProxy} \code{htmlwidget} object.
}
\description{
Unload data to the chart with proxy
}
