% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-package.R
\name{billboarder-exports}
\alias{billboarder-exports}
\alias{\%>\%}
\alias{JS}
\alias{aes}
\title{billboarder exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the billboarder package to avoid listing the magrittr
as Depends of billboarder
}
