% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-package.R
\docType{package}
\name{billboarder-package}
\alias{billboarder-package}
\title{An \code{htmlwidget} interface to the
billboard.js javascript chart library}
\description{
This package allow you to use billboard.js (\url{https://naver.github.io/billboard.js/}),
a re-usable easy interface JavaScript chart library, based on D3 v4+.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dreamRs/billboarder}
  \item \url{https://dreamrs.github.io/billboarder/}
  \item Report bugs at \url{https://github.com/dreamRs/billboarder/issues}
}

}
\author{
Victor Perrier (@dreamRs_fr)
}
