% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder.R
\name{billboarder}
\alias{billboarder}
\title{Create a Billboard.js widget}
\usage{
billboarder(
  bb_opts = list(),
  data = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{bb_opts}{A `list` in JSON format with chart parameters, see
<https://naver.github.io/billboard.js/demo/>.}

\item{data}{A `data.frame`.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Create an interactive visualization with Javascript library Billboard.js.
}
