% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-data.R
\docType{data}
\name{equilibre_mensuel}
\alias{equilibre_mensuel}
\title{Monthly supply / demand balance (january 2007 to june 2017)}
\format{
A data frame with 126 rows and 5 variables:
\describe{
  \item{date}{Date}
  \item{solde}{Supply/demand balance (in GWh)}
  \item{production}{Generation (in GWh)}
  \item{pompage}{Pumping for hydraulic generation (in GWh)}
  \item{consommation}{Consumption (in GWh)}
}
}
\source{
RTE (\url{https://odre.opendatasoft.com/explore/dataset/equilibre-national-mensuel-prod-conso-brute/})
}
\usage{
equilibre_mensuel
}
\description{
Monthly history of supply/demand balance (GWh) based on gross consumption, 
the balance of physical exchanges with foreign countries and offtakes due to pumping.
Last update : 2017-07-27.
}
\keyword{datasets}
