% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prefix}
\alias{prefix}
\title{Shortcut to add a prefix value to axis labels}
\usage{
prefix(x)
}
\arguments{
\item{x}{A character of length one.}
}
\description{
Shortcut to add a prefix value to axis labels
}
\seealso{
suffix
}
