% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-data.R
\docType{data}
\name{prod_filiere_long}
\alias{prod_filiere_long}
\title{French electricity generation by year and branch.}
\format{
A data frame with 45 rows and 3 variables:
\describe{
  \item{annee}{Year}
  \item{branche}{Source of production}
  \item{prod}{Production in TWh}
}
}
\source{
RTE (\url{https://odre.opendatasoft.com/explore/dataset/prod-national-annuel-filiere/})
}
\usage{
prod_filiere_long
}
\description{
Annual French electricity production (TWh) by branch. Last update : 2017-02-15.
}
\keyword{datasets}
