% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_example.R
\name{proxy_example}
\alias{proxy_example}
\title{Proxy use example}
\usage{
proxy_example(chart = c("gauge", "pie", "bar", "line"))
}
\arguments{
\item{chart}{Chart type for which to see an example.}
}
\description{
Launch an example app to demonstrate how to use proxy method from \code{billboarder}.
}
\examples{
\dontrun{

if (interactive()) {

proxy_examples("gauge")

}

}
}
