% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suffix}
\alias{suffix}
\title{Shortcut to add a suffix value to axis labels}
\usage{
suffix(x)
}
\arguments{
\item{x}{A character of length one.}
}
\description{
Shortcut to add a suffix value to axis labels
}
\seealso{
prefix
}
