% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_area}
\alias{bb_area}
\title{Area property for a Billboard.js chart}
\usage{
bb_area(bb, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{See \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.area}}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Area property for a Billboard.js chart
}
