% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_bar}
\alias{bb_bar}
\title{Create a bar chart}
\usage{
bb_bar(bb, data, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{data}{A \code{data.frame}, the first column will be used for x axis unless
specified otherwise in \code{...}}

\item{...}{Arguments for slot data and bar}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Create a bar chart
}
\examples{
\dontrun{

}
}
