% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_bar}
\alias{bb_bar}
\title{Helper for creating a bar chart}
\usage{
bb_bar(bb, data, stacked = FALSE, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{data}{A \code{data.frame}, the first column will be used for x axis unless
specified otherwise in \code{...}}

\item{stacked}{Logical, if several columns provided, produce a stacked bar chart, else
a dodge bar chart.}

\item{...}{Arguments for slot bar, see \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.bar}.}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating a bar chart
}
\examples{
library("billboarder")

stars <- data.frame(
  package = c("billboarder", "ggiraph", "officer", "shinyWidgets", "visNetwork"),
  stars = c(1, 176, 42, 40, 166)
)

billboarder() \%>\%
  bb_bar(data = stars)

billboarder() \%>\%
  bb_bar(data = stars, labels = TRUE) \%>\%
  bb_data(names = list(stars = "Number of stars")) \%>\% 
  bb_axis(rotated = TRUE)
}
