% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_color}
\alias{bb_color}
\title{Color property for a Billboard.js chart}
\usage{
bb_color(bb, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{See \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.color}}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Color property for a Billboard.js chart
}
\examples{
\dontrun{
# change color

}
}
