% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_export}
\alias{bb_export}
\title{Export a Billboard to PNG}
\usage{
bb_export(bb, filename = NULL)
}
\arguments{
\item{bb}{A \code{\link{billboarder}} \code{htmlwidget} object
or a \code{\link{billboarderProxy}} \code{htmlwidget} object.}

\item{filename}{A string of the filename, excluding extension (will be \code{.png}).}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Export a Billboard to PNG
}
\examples{
# TODO
}
