% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-shiny.R
\name{bb_export}
\alias{bb_export}
\title{Export a Billboard object}
\usage{
bb_export(proxy, path = NULL)
}
\arguments{
\item{proxy}{A \code{billboardProxy} \code{htmlwidget} object.}

\item{path}{Path where to save the widget.}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Export a Billboard object
}
\examples{
# TODO
}
