% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_histogram}
\alias{bb_histogram}
\title{Helper for creating an histogram}
\usage{
bb_histogram(bb, x, breaks = "Sturges", ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{x}{A numeric \code{vector}.}

\item{breaks}{Arguments passed to \code{hist}.}

\item{...}{Arguments for slot}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating an histogram
}
