% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_interaction}
\alias{bb_interaction}
\title{Interaction property for a Billboard.js chart}
\usage{
bb_interaction(bb, ...)
}
\arguments{
\item{bb}{A `billboard` `htmlwidget` object.}

\item{...}{See <https://naver.github.io/billboard.js/release/latest/doc/Options.html#.interaction>}
}
\value{
A `billboard` `htmlwidget` object.
}
\description{
Interaction property for a Billboard.js chart
}
\examples{
# Disable interactions
billboarder() \%>\%
  bb_linechart(data = c(1, 3, 2, 5, 4)) \%>\%
  bb_interaction(enabled = FALSE)

# Only mouse input
billboarder() \%>\%
  bb_linechart(data = c(1, 3, 2, 5, 4)) \%>\%
  bb_interaction(inputType = list(touch = FALSE))
}
