% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_legend}
\alias{bb_legend}
\title{Add legend parameters}
\usage{
bb_legend(bb, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://naver.github.io/billboard.js/demo/}.}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Add legend parameters
}
