% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_linechart}
\alias{bb_linechart}
\title{Helper for creating a line chart}
\usage{
bb_linechart(bb, data, type = "line", show_point = FALSE, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} or a \code{vector}.}

\item{type}{Type of chart : line, spline, step, area, area-spline, area-step.}

\item{show_point}{Whether to show each point in line.}

\item{...}{Not used}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating a line chart
}
