% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_linechart}
\alias{bb_linechart}
\title{Helper for creating a line chart}
\usage{
bb_linechart(bb, data, type = "line", show_point = FALSE, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} or a \code{vector}.}

\item{type}{Type of chart : line, spline, step, area, area-spline, area-step.}

\item{show_point}{Whether to show each point in line.}

\item{...}{Not used}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating a line chart
}
\examples{

# Different types
x <- round(rnorm(20), 2)

billboarder() \%>\% 
  bb_linechart(data = x)

billboarder() \%>\% 
  bb_linechart(data = x, type = "spline")

billboarder() \%>\% 
  bb_linechart(data = x, type = "area")

billboarder() \%>\% 
  bb_linechart(data = x, type = "area-spline")
  
  
# Timeserie with date (Date)
data("economics", package = "ggplot2")

billboarder() \%>\%
  bb_linechart(data = economics[, c("date", "psavert")]) \%>\% 
  bb_x_axis(tick = list(format = "\%Y-\%m", fit = FALSE)) \%>\%
  bb_y_axis(tick = list(format = suffix("\%")), 
            label = list(text = "Personal savings rate")) \%>\% 
  bb_legend(show = FALSE) \%>\% 
  bb_x_grid(show = TRUE) \%>\% 
  bb_y_grid(show = TRUE) \%>\% 
  bb_subchart(show = TRUE)
  

# Timeserie with datetime (POSIXct)
data("cdc_prod_filiere")

billboarder() \%>\% 
  bb_linechart(data = cdc_prod_filiere[, c("date_heure", "prod_eolien")])
  
}
