% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_lollipop}
\alias{bb_lollipop}
\title{Helper for creating a lollipop chart}
\usage{
bb_lollipop(bb, data, x = NULL, y = NULL, rotated = FALSE,
  point_color = "#112446", point_size = 8, line_color = "#000", ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{data}{A \code{data.frame}, the first column will be used for x axis unless
argument \code{x} is speciefied, the second one will be use as y values.
 If not a \code{data.frame}, an object coercible to \code{data.frame}.}

\item{x}{Character, the variable to use for the x axis.}

\item{y}{Character, the variable containing the values to map on the chart.}

\item{rotated}{Switch x and y axis position.}

\item{point_color}{Color of the lollipop.}

\item{point_size}{Size of the lollipop.}

\item{line_color}{Color of the lines between the axis and the lollipop.}

\item{...}{}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating a lollipop chart
}
\examples{
# TODO
}
