% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_padding}
\alias{bb_padding}
\title{The padding of the chart element.}
\usage{
bb_padding(bb, ...)
}
\arguments{
\item{bb}{A \code{\link{billboarder}} \code{htmlwidget} object
or a \code{\link{billboarderProxy}} \code{htmlwidget} object.}

\item{...}{See \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.padding} for possible options.}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
The padding of the chart element.
}
