% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_pie}
\alias{bb_pie}
\title{Pie property for a Billboard.js chart}
\usage{
bb_pie(bb, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{See \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.pie}}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Pie property for a Billboard.js chart
}
