% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_regions}
\alias{bb_regions}
\title{Regions property for a Billboard.js chart}
\usage{
bb_regions(bb, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{See \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.regions}}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Regions property for a Billboard.js chart
}
