% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_helpers.R
\name{bb_scatterplot}
\alias{bb_scatterplot}
\title{Helper for creating a scatter chart}
\usage{
bb_scatterplot(bb, data, x = NULL, y = NULL, group = NULL, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{data}{A \code{data.frame}}

\item{x}{Variable to map to the x-axis, if \code{NULL} first variable is used.}

\item{y}{Variable to map to the y-axis, if \code{NULL} second variable is used.}

\item{group}{Variable to use to plot data by group.}

\item{...}{unused}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
Helper for creating a scatter chart
}
\examples{
\dontrun{
billboarder() \%>\% 
  bb_scatterplot(data = iris, x = "Sepal.Length", y = "Sepal.Width")
}
}
