% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_utils.R
\name{bb_svg}
\alias{bb_svg}
\title{SVG property for a Billboard.js chart}
\usage{
bb_svg(bb, ...)
}
\arguments{
\item{bb}{A \code{billboard} \code{htmlwidget} object.}

\item{...}{See \url{https://naver.github.io/billboard.js/release/latest/doc/Options.html#.svg}}
}
\value{
A \code{billboard} \code{htmlwidget} object.
}
\description{
SVG property for a Billboard.js chart
}
