% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboarder-data.R
\docType{data}
\name{cdc_prod_filiere}
\alias{cdc_prod_filiere}
\title{French electricity generation by power source for the day of 2017-06-12.}
\format{
A data frame with 48 rows and 11 variables:
\describe{
  \item{date_heure}{Timestamp (POSIXct)}
  \item{prod_total}{Total production in MW (thermal + hydro + nuclear + solar + wind + bioenergy)}
  \item{prod_gaz}{Gas production in MW}
  \item{prod_bioenergies}{Bioenergy production in MW}
  \item{prod_hydraulique}{Hydraulic production in MW}
  \item{prod_thermique_fossile}{Fossil thermal production in MW}
  \item{prod_charbon}{Coal production in MW}
  \item{prod_eolien}{Wind production in MW}  
  \item{prod_solaire}{Solar production in MW}   
  \item{prod_nucleaire}{Nuclear production in MW}
  \item{prod_fioul}{Oil production in MW}
}
}
\source{
RTE (\url{https://opendata.reseaux-energies.fr/explore/dataset/production-quotidienne-filiere})
}
\usage{
cdc_prod_filiere
}
\description{
Average power generation (MW) per 30-minute interval within continental France, aggregated by broad power source. Last update : 2017-07-27.
}
\keyword{datasets}
